WIP 2: Refactor more into components
