+++ gtkcord/components/typing/typing.go
