+++ gtkcord/components/window/header.go
