+++ gtkcord/md/md.go
@@ -8,7 +8,6 @@ import (
 	"github.com/alecthomas/chroma"
 	"github.com/alecthomas/chroma/styles"
 	"github.com/diamondburned/arikawa/discord"
-	"github.com/diamondburned/arikawa/state"
 	"github.com/diamondburned/gtkcord3/gtkcord/semaphore"
 	"github.com/diamondburned/gtkcord3/log"
 	"github.com/gotk3/gotk3/gdk"
@@ -42,38 +41,24 @@ var (
 	fmtter   = Formatter{}
 	css      = map[chroma.TokenType]string{}
 	lexerMap = sync.Map{}
+	pool     = newPool(p)
 )
 
-type Parser struct {
-	pool  sync.Pool
-	State *state.State
+var ChannelPressed func(ev *gdk.EventButton, ch discord.Channel)
+var UserPressed func(ev *gdk.EventButton, user discord.GuildUser)
 
-	ChannelPressed func(ev *gdk.EventButton, ch discord.Channel)
-	UserPressed    func(ev *gdk.EventButton, user discord.GuildUser)
-}
-
-func NewParser(s *state.State) *Parser {
+func init() {
 	log.Debugln("REGEX:", strings.Join(regexes, "|"))
 
+	style = styles.Get(HighlightStyle)
 	if style == nil {
-		style = styles.Get(HighlightStyle)
-		if style == nil {
-			panic("Unknown highlighting style: " + HighlightStyle)
-		}
-
-		css = styleToCSS(style)
+		panic("Unknown highlighting style: " + HighlightStyle)
 	}
-
-	p := &Parser{
-		State: s,
-	}
-	p.pool = newPool(p)
-
-	return p
+	css = styleToCSS(style)
 }
 
-func (p *Parser) Parse(md []byte, buf *gtk.TextBuffer) {
-	p.ParseMessage(nil, nil, md, buf)
+func Parse(md []byte, buf *gtk.TextBuffer) {
+	ParseMessage(nil, nil, md, buf)
 }
 
 type Discord interface {
@@ -81,8 +66,8 @@ type Discord interface {
 	Member(guild, user discord.Snowflake) (*discord.Member, error)
 }
 
-func (p *Parser) ParseMessage(d Discord, m *discord.Message, md []byte, buf *gtk.TextBuffer) {
-	s := p.pool.Get().(*mdState)
+func ParseMessage(d Discord, m *discord.Message, md []byte, buf *gtk.TextBuffer) {
+	s := pool.Get().(*mdState)
 	s.use(buf, md, d, m)
 
 	var tree func(i int)
@@ -130,5 +115,5 @@ func (p *Parser) ParseMessage(d Discord, m *discord.Message, md []byte, buf *gtk
 	s.attr = 0
 	s.color = ""
 
-	p.pool.Put(s)
+	pool.Put(s)
 }
