+++ gtkcord/ningen/channel.go
@@ -1,0 +1,1 @@
+package ningen
